Fix: 모집글 게시판 모집 마감 나타나지 않는 버그 수정(##264)
