+++ src/main/frontend/src/Routes/PostAddForm/components/color.js
@@ -1,0 +1,2 @@
+export const MainBLUE = "bg-blue-200";
+export const LightMainBLUE = "bg-blue-100";
